Fix several casting warnings about longs

+++ TwitterStats/TwitterStats/EmojiTableViewController.m
@@ -39,7 +39,7 @@
     cell.textLabel.font = [UIFont fontWithName:@"Helvetica Neue" size:17.0];
     
     NSInteger rowNumber = indexPath.row + 1;
-    cell.textLabel.text = [NSString stringWithFormat:@"%ld. %@ %@ count: %lld", rowNumber, emojiImageString, emojiName, emoji.count];
+    cell.textLabel.text = [NSString stringWithFormat:@"%ld. %@ %@ count: %lld", (long)rowNumber, emojiImageString, emojiName, emoji.count];
 }
 
 @end
